requirements.txt
biopython
pandas
numpy
matplotlib
seaborn
scikit-learn
